requirements.txt
cycler==0.10.0
decorator==4.4.2
dill==0.3.3
dominate==2.6.0
imageio==2.9.0
imgaug==0.4.0
kiwisolver==1.3.1
lpips==0.1.3
networkx==2.5.1
opencv-python==4.5.2.52
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
Shapely==1.7.1
six==1.16.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.61.0
typing-extensions==3.10.0.0
